String of hexadecimal digits to set linetype.
